third_party/blink/web_tests/wpt_internal/css/css-conditional/container-queries/orthogonal-wm-container-query.html

<!doctype html>
<title>Orthogonal writing-mode change in @container</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="stylesheet" href="/fonts/ahem.css">
<style>
  #container {
    contain: size layout style;
    width: 50vw;
    height: 50vh;
  }
  #orthogonal {
    font: 50px/1 Ahem;
  }
  @container (max-width: 100px) {
    #orthogonal {
      writing-mode: vertical-lr;
    }
  }
</style>
<div id="container">
  <div id="orthogonal">XX</div>
</div>
<script>
  test(() => {
    assert_equals(orthogonal.offsetWidth, container.offsetWidth);
  }, "Initial non-orthogonal width");

  test(() => {
    container.style.width = "100px";
    assert_equals(orthogonal.offsetWidth, 50);
    assert_not_equals(orthogonal.offsetWidth, container.offsetWidth);
  }, "Orthogonal width");
</script>
