third_party/blink/web_tests/flag-specific/highdpi/virtual/container-queries/wpt_internal/css/css-conditional/container-queries/orthogonal-wm-container-query-expected.txt

This is a testharness.js-based test.
PASS Initial non-orthogonal width
FAIL Orthogonal width assert_equals: expected 50 but got 100
Harness: the test ran to completion.
